refactor: extract ExCS preview seeds to fix module length
Move the public Scratch preview seed helpers out of SeedsHelper so
RuboCop Metrics/ModuleLength stays under the limit.
